upper-case aliases if they're identical to the protocol name
This addresses PR44311 (11 years later :-), which was never completely fixed,
as the regeneration of the protocols files reinstated the lowercased (identical)
alias for the "manet" protocol.  As a result, when importing the protocols file,
one had to know that one needed to manually update the entry so as not to break
things.