Add print/poppler-qt6 version 22.10.0 to pkgsrc-current
This is the QT6 wrapper for poppler, a PDF rendering library. It
is a fork of the xpdf 3.0 PDF viewer developed by Derek Noonburg
of Glyph and Cog, LLC.

Poppler provides PDF rendering functionality as a shared library,
to centralize the maintenence effort and to reduce duplicated
effort. And the poppler developers would like to move libpoppler
forward in a number of areas that doesn't fit within the goals of
xpdf.