x11/xlax: import package from wip.
xlax is an X11 input multiplexer: keyboard input can be set from a xlax
instance to a group of windows simultaneously.
It allows the user to assign a custom string to each window (e.g. a
hostname) that can be sent with a button click.

An included script handles starting up and arranging windows, and
permits to define commonly used window groupings in a config file.